// src/lib/types.ts
export type StoreSettings = {
  storeName: string;
  storeCity: string;
  storeAddress: string;
  pixKey: string;
  storePhone: string;
  logoUrl?: string;
  accessControlEnabled?: boolean;
  commercialHourStart?: string;
  commercialHourEnd?: string;
  wapiInstance?: string;
  wapiToken?: string;
};